Update ExchangebetweenDB.sql
This script was written to cover a case (In my workplace ) when a new DB needs to be installed on the customer site instead of the old one (for some reason, defects\faults etc.. ) and the customer has a large amount of structural data (organization\departments\subscriber\cars) which he does not want to type again. 
copying the data from a backup is not relevant due to GUIDS related to each object.
The process retreives the data from the source and use other solution i have written to create and inject the records to the target DB . 

Feel free to adopt any part of it for your needs.